Implementation of FaaastRegistryHandler
